+++ FoundationDB.Client/FdbTransaction.Snapshot.cs
@@ -82,7 +82,7 @@ namespace FoundationDB.Client
 			IFdbTransactionOptions IFdbReadOnlyTransaction.Options => m_parent.Options;
 
 			/// <inheritdoc />
-			public (int Keys, int Size) GetReadStatistics() => m_parent.GetReadStatistics();
+			public (int Keys, long Size) GetReadStatistics() => m_parent.GetReadStatistics();
 
 			/// <inheritdoc />
 			public void EnsureCanRead()
